A little refactoring ;)
