+++ x-pack/plugins/endpoint/common/types.ts
@@ -23,6 +23,7 @@ export type ImmutableSet<T> = ReadonlySet<Immutable<T>>;
 export type ImmutableObject<T> = { readonly [K in keyof T]: Immutable<T[K]> };
 
 export class EndpointAppConstants {
+  static BASE_API_URL = '/api/endpoint';
   static ALERT_INDEX_NAME = 'my-index';
   static ENDPOINT_INDEX_NAME = 'endpoint-agent*';
 }
@@ -122,6 +123,18 @@ export interface EndpointMetadata {
     };
   };
 }
+
+export interface ESTotal {
+  value: number;
+  relation: string;
+}
+
+export interface AlertDataWrapper {
+  _id: string;
+  _source: AlertData;
+}
+
+export type AlertHits = AlertDataWrapper[];
 
 /**
  * The PageId type is used for the payload when firing userNavigatedToPage actions
