+++ x-pack/plugins/endpoint/server/routes/alerts/details/schemas.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+import { schema } from '@kbn/config-schema';
+
+export const alertDetailsReqSchema = schema.object({
+  id: schema.string(),
+});
